Make sure reddit-v2-events new enough for thrift-unofficial (#11)
Before this version, reddit-v2-events depended on "thrift" directly. As
of Baseplate.py v2.0 we depend on thrift-unofficial since we're ahead of
the official releases on PyPI. The two packages overwrite eachother so
this conflicting requirement was breaking usage of Thrift in
Baseplate.py. This ensures projects are using a sufficiently new
reddit-v2-events to not get the conflicting installation.